pyproject.toml
[project]
name = "timescale-hybrid-search-demo"
version = "1.0.0"
description = "Interactive hybrid search demo with TimescaleDB, pgvectorscale, and PostgreSQL full-text search"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Timescale", email = "support@timescale.com" }
]
keywords = ["timescaledb", "pgvector", "hybrid-search", "vector-search", "full-text-search", "rag"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Database connectivity
    "psycopg[binary]>=3.1.0",

    # Environment management
    "python-dotenv>=1.0.0",

    # Terminal UI
    "rich>=13.0.0",

    # Embedding generation (required for interactive query input)
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/timescale/timescale-k8s-hybrid-search"
Documentation = "https://docs.timescale.com/"
Repository = "https://github.com/timescale/timescale-k8s-hybrid-search"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
# Cross-platform support: Works on Linux x86, macOS Apple Silicon, and other platforms
# uv automatically handles platform-specific binary wheels

[dependency-groups]
# Development dependencies (optional, for maintainers)
dev = [
    "ipython>=8.0.0",
    "pytest>=7.0.0",
]

[tool.uv.sources]
# Use PyPI for all packages (cross-platform wheels available)
# torch and sentence-transformers have pre-built wheels for:
# - Linux x86_64
# - macOS x86_64 (Intel)
# - macOS arm64 (Apple Silicon)
# - Windows x86_64
